+++ src/test/tests/issuesDataProvider.test.ts
@@ -3,7 +3,6 @@ import { faker } from '@faker-js/faker';
 import * as vscode from 'vscode';
 import { ScanCacheManager } from '../../main/cache/scanCacheManager';
 import { DependenciesTreeNode } from '../../main/treeDataProviders/dependenciesTree/dependenciesTreeNode';
-import { IssuesDataProvider, LicensesTitleNode, VulnerabilitiesTitleNode, VulnerabilityNode } from '../../main/treeDataProviders/issuesDataProvider';
 import { GeneralInfo } from '../../main/types/generalInfo';
 import { IIssueCacheObject } from '../../main/types/issueCacheObject';
 import { IIssueKey } from '../../main/types/issueKey';
@@ -12,8 +11,9 @@ import * as issueSeverity from '../../main/types/severity';
 import { Severity } from '../../main/types/severity';
 import { TestMemento } from './utils/testMemento.test';
 import { ScanUtils } from '../../main/utils/scanUtils';
-import { DependencyDetailsProvider } from '../../main/treeDataProviders/dependencyDetailsProvider';
+import { DependencyDetailsProvider } from '../../main/treeDataProviders/dependenciesTree/dependencyDetailsProvider';
 import { PackageType } from '../../main/types/projectType';
+import { IssuesDataProvider, LicensesTitleNode, VulnerabilitiesTitleNode, VulnerabilityNode } from '../../main/treeDataProviders/dependenciesTree/issuesDataProvider';
 // import { SourceCodeTreeDataProvider } from '../../main/treeDataProviders/sourceCodeTree/sourceCodeTreeDataProvider';
 // import { TreesManager } from '../../main/treeDataProviders/treesManager';
 // import { ConnectionManager } from '../../main/connect/connectionManager';
